Split associations and singleton associations
Previously a class was marked singleton by including the SingleResource module. However this came with the huge downside that a class was either singleton or not. It was not possible for the same class to be a singleton resource in one context, and a many resource in another. This commit fixes this problem, by splitting the association systen into singleton and standard association. The singleton-ness of an object is thus no longer defined by it’s class, but by the individual association.